Feby Maulana

+++ routes/web.php
@@ -4,6 +4,8 @@ use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\PelatihController;
 use App\Http\Controllers\AdminController;
 use App\Http\Controllers\MuridController;
+use App\Http\Controllers\LoginController;
+
 /*
 |--------------------------------------------------------------------------
 | Web Routes
@@ -12,7 +14,7 @@ use App\Http\Controllers\MuridController;
 | Here is where you can register web routes for your application. These
 | routes are loaded by the RouteServiceProvider within a group which
 | contains the "web" middleware group. Now create something great!
-|
+|   
 */
 
 Route::get('/', function () {
@@ -25,41 +27,48 @@ Route::get("/layouts", function() {
 
 Route::get('/dashboard', function () {
     return view('/admin/dashboard');
-});
+})->middleware('admin');
+
+// login adminController
+Route::get('/logout', [loginController::class, "logout"]);
+
+Route::get("/login", [LoginController::class, "login"])->middleware('guest');
+
+Route::POST("/post_login", [LoginController::class, "post_login"] );
 
 // Data Pelatih
-Route::get('/pelatih', [PelatihController::class, 'index']);
-Route::get("/pelatih/edit/{id}", [PelatihController::class, 'edit']);
-Route::get('/pelatih/hapus/{id}', [PelatihController::class, 'destroy']);
-Route::post('/pelatih/update', [PelatihController::class, 'update']);
+Route::get('/pelatih', [PelatihController::class, 'index'])->middleware('admin');
+Route::get("/pelatih/edit/{id}", [PelatihController::class, 'edit'])->middleware('admin');
+Route::get('/pelatih/hapus/{id}', [PelatihController::class, 'destroy'])->middleware('admin');
+Route::post('/pelatih/update', [PelatihController::class, 'update'])->middleware('admin');
 
-Route::post('/pelatih/add', [PelatihController::class, 'store']);
+Route::post('/pelatih/add', [PelatihController::class, 'store'])->middleware('admin');
 
 Route::get('pelatih/addpelatih', function () {
 
-    return view('/admin/pelatih/addpelatih');
+    return view('/admin/pelatih/addpelatih')->middleware('admin');
 });
 
 Route::get('/murid', function () {
-    return view('/admin/murid');
+    return view('/admin/murid')->middleware('admin');
 });
 
-Route::get("/admin", [AdminController::class, "dashboard"] );
-Route::delete('{pelatih}/delete','PelatiCotroller@destroy')->name('pelatih.destroy');
+Route::get("/admin", [AdminController::class, "dashboard"] )->middleware('admin');
+Route::delete('{pelatih}/delete','PelatiCotroller@destroy')->name('pelatih.destroy')->middleware('admin');
 
 
 //Data Murid
-Route::get('/murid', [MuridController::class, 'index']);
-Route::get("/murid/edit/{id}", [MuridController::class, 'edit']);
-Route::get('/murid/hapus/{id}', [MuridController::class, 'destroy']);
-Route::post('/murid/update', [MuridController::class, 'update']);
+Route::get('/murid', [MuridController::class, 'index'])->middleware('admin');
+Route::get("/murid/edit/{id}", [MuridController::class, 'edit'])->middleware('admin');
+Route::get('/murid/hapus/{id}', [MuridController::class, 'destroy'])->middleware('admin');
+Route::post('/murid/update', [MuridController::class, 'update'])->middleware('admin');
 
-Route::post('murid/add', [MuridController::class, 'store']);
+Route::post('murid/add', [MuridController::class, 'store'])->middleware('admin');
 
 Route::get('murid/addmurid', function () {
 
-    return view('/admin/murid/addmurid');
+    return view('/admin/murid/addmurid')->middleware('admin');
 });
 
-Route::get("/admin", [AdminController::class, "dashboard"] );
-Route::delete('{murid}/delete','MuridCotroller@destroy')->name('murid.destroy');
+Route::get("/admin", [AdminController::class, "dashboard"] )->middleware('admin');
+Route::delete('{murid}/delete','MuridCotroller@destroy')->name('murid.destroy')->middleware('admin');
